Fixing Graphing for comparison among Energy Consumption
The eigen set test for a unique type of test has N iterations, so the
perf for x86 collected the whole set for a unique type, and the
calculations for x86 was not fair. Powerpc was closed, the collected
samples are related to the whole set as x86 but the calculations was
using the real_time of one test, so the whole set was choice and more
sensors datada was collected to have an average as well.